Added Scaffolding To Add New Fonts
Added a bunch o' js that will map out a smaller Font Definition to its full expanded coordinates so that I don't have to hand type everything. It's not finished and needs to be expanded for Fonts with both Upper and Lowercase values.

updated the basic landing page with credit and references to the original authors, and will now begin the process of adding additional controls to assist with Font Selection and Special Glyph / Character Selection as certain fonts contain more than just alpha-numeric & standard special char values.